+++ lib/flutter_slidable_list_view.dart
@@ -3,3 +3,4 @@ library flutter_slidable_list_view;
 export 'package:flutter_slidable_list_view/widgets/slide_list.dart';
 export 'package:flutter_slidable_list_view/support/base_def.dart';
 export 'package:flutter_slidable_list_view/widgets/action_widgets.dart';
+export 'package:flutter_slidable_list_view/widgets/slide_item.dart';