+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## [1.0.0] - change description
 ## [1.0.1] - update description
 ## [1.2.0] - finish the description of it,did't change main code 
+## [1.2.1] - add itemCount param,you can set the item count by yourself 
+## [1.2.2] - fix v1.2.1 item count bug
+## [1.2.3] - fix crash when remove item
+## [1.2.4] - add UnSlidableWidget to make some item not slide when touch 
+## [1.2.5] - when dataList'length change, list will jump to top <-- bug fixed
+## [1.2.6] - support listView padding